refactor: move header to the separate folder
